Some refactoring FormValidationTest
